requirements.txt
Crypto
flask
python-dotenv
